requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
Django==1.11.20
django-cors-headers==2.4.0
django-extensions==2.1.5
djangorestframework==3.9.1
future==0.17.1
geocoder==1.38.1
idna==2.8
pkg-resources==0.0.0
psycopg2==2.7.7
pytz==2018.9
ratelim==0.1.6
requests==2.21.0
six==1.12.0
typing==3.6.6
urllib3==1.24.1
